checkout now is the first operation on mainClosure()
